chore(utxo-lib): removed functions as class methods and became util func
TICKET: BTC-2183
