+++ modules/utxo-lib/src/bitgo/index.ts
@@ -21,6 +21,7 @@ export * from './zcash';
 export * from './tnumber';
 export * from './litecoin';
 export * from './PsbtUtil';
+export * from './ProprietaryKeyValUtils';
 
 import { PsbtInput } from 'bip174/src/lib/interfaces';
 /**
